+++ CHANGES.rst
@@ -20,6 +20,9 @@ Changelog
 2.3.2 (2024-01-03)
 ------------------
 
+- Add creation and modification dates to serialized booking
+  [folix-01]
+ 
 - Add fullobjects in @bookings
   [mamico]